Move testing/e2e into src
