+++ src/webhook/mutation/pod_mutator_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/Dynatrace/dynatrace-operator/src/mapper"
 	"github.com/Dynatrace/dynatrace-operator/src/scheme"
 	"github.com/Dynatrace/dynatrace-operator/src/scheme/fake"
+	t_utils "github.com/Dynatrace/dynatrace-operator/src/testing"
 	dtwebhook "github.com/Dynatrace/dynatrace-operator/src/webhook"
-	t_utils "github.com/Dynatrace/dynatrace-operator/testing"
 	jsonpatch "github.com/evanphx/json-patch"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
